+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 > Release Date: (Not Released)
 
+- Fixed `Invoke-IamDeployment` cmdlet when an isolated workload domain was deployed assigning global permissions was failing.
 - Added `Export-NsxFederationJsonSpec` cmdlet to generate a JSON specification file for NSX Federation.
 - Added `Invoke-NsxFederationDeployment` cmdlet to perform end-to-end automation of NSX Federation for VMware Cloud Foundation.
 - Added `Deploy-NsxGlobalManager` cmdlet to deploy a NSX Global Manager appliance.